# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register a user",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "username",
                                    "department_id",
                                    "phone"
                                ],
                                "properties": {
                                    "department_id": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08122223334"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/courses": {
            "get": {
                "tags": [
                    "courses"
                ],
                "description": "Returns courses",
                "operationId": "findCourses",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sorts items base of field provided",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "-id",
                                "created_at",
                                "-created_at",
                                "code",
                                "-code",
                                "-name",
                                "name"
                            ]
                        },
                        "examples": {
                            "-id": {
                                "summary": "sorts in an descending manner by append a '-' ",
                                "value": "-id"
                            },
                            "id": {
                                "summary": "sorts in an ascending manner",
                                "value": "id"
                            }
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filters base on the parameter passed e.g filter[field]=value",
                        "required": false,
                        "style": "deepObject",
                        "schema": {
                            "properties": {
                                "level_id": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "code": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "mat101"
                                    }
                                },
                                "name": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Elementary Algebra"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "includes the details of the field included e.g include='materials'",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "department",
                                    "level",
                                    "materials",
                                    "quizzes",
                                    "lectures"
                                ]
                            }
                        }
                    },
                    {
                        "$ref": "#components/parameters/range"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    }
                }
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "courses"
                ],
                "description": "Returns a material based on a single ID",
                "operationId": "findCoursesById",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        },
        "/api/courses/{id}/departments": {
            "get": {
                "tags": [
                    "materials"
                ],
                "description": "Returns a material based on a single ID",
                "operationId": "findCourseDepartmentsById",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Department"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Course": {
                "required": [
                    "code",
                    "level_id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "the course title or course name",
                        "type": "string"
                    },
                    "code": {
                        "description": "the course code",
                        "type": "string"
                    },
                    "level_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Course"
                }
            },
...
            "Department": {
                "required": [
                    "department",
                    "faculty_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "department": {
                        "type": "string"
                    },
                    "banner": {
                        "type": "string"
                    },
                    "faculty_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Department"
                }
            },
...
            "User": {
                "required": [
                    "password",
                    "department_id",
                    "email",
                    "username",
                    "phone_number",
                    "authorization_level"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "department_id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "username": {
                        "description": "User unique username",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "User phone number",
                        "type": "string"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "email_verified_at": {
                        "description": "Datetime marker of verification status",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2019-02-25 12:59:20"
                    },
                    "authorization_level": {
                        "description": "number from 1-11 level 5 above = moderator, level 8 above is admin",
                        "type": "integer"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "responses": {
            "404": {
                "description": "not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "400": {
                "description": "Error: Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Requested filter(s) `description, are not allowed. Allowed filter(s) are `created_at` "
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
...
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID of model to fetch",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        }
    },
    "security": [
        []
    ]
}